renamed service module to routes
